# swagger/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Internal API",
        "title": "Bigfood",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8000",
    "basePath": "/",
    "paths": {
...
        "/table/delete-all": {
            "delete": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Delete all tables",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "table"
                ],
                "summary": "Delete all tables",
                "parameters": [
                    {
                        "description": "Body",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/tableDeleteAll.Message"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    },
                    "400": {
                        "description": "Invalid data",
                        "schema": {
                            "$ref": "#/definitions/server.ResponseError"
                        }
                    },
                    "401": {
                        "description": "Access Denied",
                        "schema": {
                            "$ref": "#/definitions/server.ResponseError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/server.ResponseError"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "server.ResponseError": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string",
                    "example": "Error message"
                }
            }
        },
...
        "tableDeleteAll.Message": {
            "type": "object",
            "required": [
                "cafe-id"
            ],
            "properties": {
                "cafe-id": {
                    "type": "string",
                    "example": "uuid"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}